+++ abstraction/target/python/guilder_abstraction.py
@@ -1,56 +1,56 @@
 from abc import ABC, abstractmethod
 
-"""test server network connection"""
 class TestServer(ABC):
-	"""test ping"""
+	"""test server network connection"""
 	@abstractmethod
 	def ping(self) -> bool:
+		"""test ping"""
 		pass
 
-	"""get server local time"""
 	@abstractmethod
 	def get_server_time(self) -> int:
+		"""get server local time"""
 		pass
 
 
-"""get market data such as symbol, price and volume"""
 class GetMarketData(ABC):
-	"""get symbol, such as BTCUSD"""
+	"""get market data such as symbol, price and volume"""
 	@abstractmethod
 	def get_symbol(self) -> list[str]:
+		"""get symbol, such as BTCUSD"""
 		pass
 
-	"""get mid-price of a symbol (e.g. BTCUSD -> 67000.0)"""
 	@abstractmethod
 	def get_price(self, symbol: str) -> float:
+		"""get mid-price of a symbol (e.g. BTCUSD -> 67000.0)"""
 		pass
 
-	"""get orderbook"""
 	@abstractmethod
 	def get_orderbook(self, symbol: str) -> dict[float, float]:
+		"""get orderbook"""
 		pass
 
 
-"""place, change, cancel order"""
 class ManageOrder(ABC):
-	"""place order, return cloid"""
+	"""place, change, cancel order"""
 	@abstractmethod
 	def place_order(self, symbol: str, price: int, volume: int) -> int:
+		"""place order, return cloid"""
 		pass
 
-	"""change order"""
 	@abstractmethod
 	def change_order_by_cloid(self, cloid: int, price: int, volume: int) -> int:
+		"""change order"""
 		pass
 
-	"""cancel order by cloid"""
 	@abstractmethod
 	def cancel_order(self, cloid: int) -> int:
+		"""cancel order by cloid"""
 		pass
 
-	"""cancel all order regardless of cloid/symbol"""
 	@abstractmethod
 	def cancel_all_order(self) -> bool:
+		"""cancel all order regardless of cloid/symbol"""
 		pass