+++ abstraction/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to both codegen and trading.yaml will be documented in this 
 ## [0.1.2] - 2024-08-26
 ### Added
 - support description
-- 
+- support map
+
 ## [0.1.1] - 2024-08-21
 ### Added
 - support to list